Add files via upload
This project is developed to manage the blood stock in the "BLOOD BANK" and the blood packets are maintained in the database. New blood details are entered in to the project to manage blood details. Blood donor details are entered and maintained in the database. Blood sales and blood purchase are entered and maintained in this project. Blood stock reports, sales reports and blood purchase reports are managed in this project.